Update compound for v2 with new assets

+++ src/com/conlan/compound/serialization/CurrentBlockObject.java
@@ -13,7 +13,7 @@ public class CurrentBlockObject {
 	
 	public long Block() {
 		if (block_ == -1) {
-			block_ = Web3Service.ToNumeric(result.substring(2)); // offset for the initial "0x"
+			block_ = Web3Service.ToNumeric(result.substring(2)).longValue(); // offset for the initial "0x"
 		}
 		
 		return block_;
